Rename QFieldService to QFieldCloudService
